requirements.txt
anyio==4.12.1
certifi==2026.2.25
defusedxml==0.7.1
fonttools==4.61.1
fpdf2==2.8.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.2
pandas==3.0.1
pillow==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.6
six==1.17.0
typing_extensions==4.15.0
